Moved ContextStore into a module
